+++ binance/depthcache.py
@@ -1,9 +1,9 @@
 # coding=utf-8
-
 from operator import itemgetter
+import asyncio
 import time
 
-from .websockets import BinanceSocketManager
+from .streams import BinanceSocketManager
 
 
 class DepthCache(object):
@@ -120,15 +120,18 @@ class DepthCache(object):
 class BaseDepthCacheManager(object):
     DEFAULT_REFRESH = 60 * 30  # 30 minutes
 
-    def __init__(self, client, symbol, callback=None, refresh_interval=None, bm=None, limit=10):
-        """Initialise the DepthCacheManager
+    @classmethod
+    async def create(cls, client, loop, symbol, coro=None, refresh_interval=None, bm=None, limit=10):
+        """Create a DepthCacheManager instance
 
         :param client: Binance API client
         :type client: binance.Client
+        :param loop:
+        :type loop:
         :param symbol: Symbol to create depth cache for
         :type symbol: string
-        :param callback: Optional function to receive depth cache updates
-        :type callback: function
+        :param coro: Optional coroutine to receive depth cache updates
+        :type coro: async coroutine
         :param refresh_interval: Optional number of seconds between cache refresh, use 0 or None to disable
         :type refresh_interval: int
         :param bm: Optional BinanceSocketManager
@@ -137,21 +140,25 @@ class BaseDepthCacheManager(object):
         :type limit: int
 
         """
+        self = cls()
         self._client = client
+        self._loop = loop
         self._symbol = symbol
         self._limit = limit
-        self._callback = callback
+        self._coro = coro
         self._last_update_id = None
-        self._depth_message_buffer = []
         self._bm = bm
         self._refresh_interval = refresh_interval or self.DEFAULT_REFRESH
         self._conn_key = None
 
-        self._start_socket()
-        self._init_cache()
+        await asyncio.gather(
+            self._init_cache(),
+            self._start_socket(),
+        )
+        return self
 
-    def _init_cache(self):
-        """Initialise the depth cache and set a refresh time
+    async def _init_cache(self):
+        """Initialise the depth cache calling REST endpoint
 
         :return:
         """
@@ -163,22 +170,20 @@ class BaseDepthCacheManager(object):
         if self._refresh_interval:
             self._refresh_time = int(time.time()) + self._refresh_interval
 
-    def _start_socket(self):
+    async def _start_socket(self):
         """Start the depth cache socket
 
         :return:
         """
         if self._bm is None:
-            self._bm = BinanceSocketManager(self._client)
+            self._bm = BinanceSocketManager(self._client, self._loop)
 
-        self._conn_key = self._get_conn_key()
-        if not self._bm.is_alive():
-            self._bm.start()
+        self._conn_key = await self._get_conn_key()
 
-    def _get_conn_key(self):
+    async def _get_conn_key(self):
         raise NotImplementedError
 
-    def _depth_event(self, msg):
+    async def _depth_event(self, msg):
         """Handle a depth event
 
         :param msg:
@@ -188,15 +193,15 @@ class BaseDepthCacheManager(object):
 
         if 'e' in msg and msg['e'] == 'error':
             # close the socket
-            self.close()
+            await self.close()
 
             # notify the user by returning a None value
-            if self._callback:
-                self._callback(None)
+            if self._coro:
+                await self._coro(None)
 
-        self._process_depth_message(msg)
+        await self._process_depth_message(msg)
 
-    def _process_depth_message(self, msg, buffer=False):
+    async def _process_depth_message(self, msg, buffer=False):
         """Process a depth event message.
 
         :param msg: Depth event message.
@@ -205,23 +210,24 @@ class BaseDepthCacheManager(object):
         """
 
         # add any bid or ask values
-        for bid in msg['b']:
-            self._depth_cache.add_bid(bid)
-        for ask in msg['a']:
-            self._depth_cache.add_ask(ask)
-
-        # keeping update time
-        self._depth_cache.update_time = msg['E']
+        self._apply_orders(msg)
 
         # call the callback with the updated depth cache
-        if self._callback:
-            self._callback(self._depth_cache)
+        if self._coro:
+            await self._coro(self._depth_cache)
 
         # after processing event see if we need to refresh the depth cache
         if self._refresh_interval and int(time.time()) > self._refresh_time:
-            self.close()
-            self._init_cache()
-            self._start_socket()
+           await self._init_cache()
+
+    def _apply_orders(self, msg):
+        for bid in msg.get('b', []) + msg.get('bids', []):
+            self._depth_cache.add_bid(bid)
+        for ask in msg.get('a', []) + msg.get('asks', []):
+            self._depth_cache.add_ask(ask)
+
+        # keeping update time
+        self._depth_cache.update_time = msg.get('E') or msg.get('lastUpdateId')
 
     def get_depth_cache(self):
         """Get the current depth cache
@@ -231,15 +237,13 @@ class BaseDepthCacheManager(object):
         """
         return self._depth_cache
 
-    def close(self, close_socket=False):
+    async def close(self):
         """Close the open socket for this manager
 
         :return:
         """
-        self._bm.stop_socket(self._conn_key)
-        if close_socket:
-            self._bm.close()
-        time.sleep(1)
+        await self._bm.close()
+        await asyncio.sleep(1)
         self._depth_cache = None
 
     def get_symbol(self):
@@ -252,16 +256,17 @@ class BaseDepthCacheManager(object):
 
 class DepthCacheManager(BaseDepthCacheManager):
 
-    def __init__(self, client, symbol, callback=None, refresh_interval=None, bm=None, limit=500,
-                 ws_interval=None):
+    @classmethod
+    async def create(cls, client, loop, symbol, coro=None, refresh_interval=None, bm=None, limit=500, ws_interval=None):
         """Initialise the DepthCacheManager
 
         :param client: Binance API client
         :type client: binance.Client
+        :param loop: asyncio loop
         :param symbol: Symbol to create depth cache for
         :type symbol: string
-        :param callback: Optional function to receive depth cache updates
-        :type callback: function
+        :param coro: Optional function to receive depth cache updates
+        :type coro: function
         :param refresh_interval: Optional number of seconds between cache refresh, use 0 or None to disable
         :type refresh_interval: int
         :param limit: Optional number of orders to get from orderbook
@@ -270,12 +275,11 @@ class DepthCacheManager(BaseDepthCacheManager):
         :type ws_interval: int
 
         """
+        self = await super().create(client, loop, symbol, coro, refresh_interval, bm, limit)
         self._ws_interval = ws_interval
-        self._last_update_id = None
-        self._depth_message_buffer = []
-        super().__init__(client, symbol, callback, refresh_interval, bm, limit)
+        return self
 
-    def _init_cache(self):
+    async def _init_cache(self):
         """Initialise the depth cache calling REST endpoint
 
         :return:
@@ -283,12 +287,13 @@ class DepthCacheManager(BaseDepthCacheManager):
         self._last_update_id = None
         self._depth_message_buffer = []
 
-        res = self._client.get_order_book(symbol=self._symbol, limit=self._limit)
+        res = await self._client.get_order_book(symbol=self._symbol, limit=self._limit)
 
         # initialise or clear depth cache
-        super()._init_cache()
+        await super()._init_cache()
 
         # process bid and asks from the order book
+        self._apply_orders(res)
         for bid in res['bids']:
             self._depth_cache.add_bid(bid)
         for ask in res['asks']:
@@ -299,26 +304,29 @@ class DepthCacheManager(BaseDepthCacheManager):
 
         # Apply any updates from the websocket
         for msg in self._depth_message_buffer:
-            self._process_depth_message(msg, buffer=True)
+            await self._process_depth_message(msg, buffer=True)
 
         # clear the depth buffer
         self._depth_message_buffer = []
 
-    def _start_socket(self):
+    async def _start_socket(self):
         """Start the depth cache socket
 
         :return:
         """
-        super()._start_socket()
+        if not getattr(self, '_depth_message_buffer'):
+            self._depth_message_buffer = []
+
+        await super()._start_socket()
 
         # wait for some socket responses
         while not len(self._depth_message_buffer):
-            time.sleep(1)
+            await asyncio.sleep(1)
 
-    def _get_conn_key(self):
-        return self._bm.start_depth_socket(self._symbol, self._depth_event, interval=self._ws_interval)
+    async def _get_conn_key(self):
+        return await self._bm.start_depth_socket(self._symbol, self._depth_event)
 
-    def _process_depth_message(self, msg, buffer=False):
+    async def _process_depth_message(self, msg, buffer=False):
         """Process a depth event message.
 
         :param msg: Depth event message.
@@ -337,29 +345,23 @@ class DepthCacheManager(BaseDepthCacheManager):
         elif msg['U'] != self._last_update_id + 1:
             # if not buffered check we get sequential updates
             # otherwise init cache again
-            self._init_cache()
+            await self._init_cache()
 
         # add any bid or ask values
-        for bid in msg['b']:
-            self._depth_cache.add_bid(bid)
-        for ask in msg['a']:
-            self._depth_cache.add_ask(ask)
-
-        # keeping update time
-        self._depth_cache.update_time = msg['E']
+        self._apply_orders(msg)
 
         # call the callback with the updated depth cache
-        if self._callback:
-            self._callback(self._depth_cache)
+        if self._coro:
+            await self._coro(self._depth_cache)
 
         self._last_update_id = msg['u']
 
         # after processing event see if we need to refresh the depth cache
         if self._refresh_interval and int(time.time()) > self._refresh_time:
-            self._init_cache()
+            await self._init_cache()
 
 
 class OptionsDepthCacheManager(BaseDepthCacheManager):
 
-    def _get_conn_key(self):
-        return self._bm.start_options_depth_socket(self._symbol, self._depth_event)
+    async def _get_conn_key(self):
+        return await self._bm.start_options_depth_socket(self._symbol, self._depth_event)
