+++ requirements.txt
@@ -1,5 +1,4 @@
-aiohttp==2.3.10
-autobahn==21.3.1
+aiohttp==3.7.4
 certifi==2020.12.5
 chardet==3.0.4
 cryptography==3.4.7
@@ -7,6 +6,5 @@ dateparser==0.7.0
 pyOpenSSL==20.0.1
 requests==2.25.1
 service-identity==18.1.0
-Twisted==20.3.0
 ujson
-websockets==3.4
+websockets==8.1
